requirements.txt
   github-copilot-sdk==0.1.23